UICommon : make grayscale RenderScript compatible with GPU execution
The Adreno GPU RenderScript libraries do not support operations with
double-precision floating point numbers. Update the constants in
grayscale.rs to be single-precision floats so that the script may
run on the GPU.

This resolves warnings like the following seen as a result of the
GPU RenderScript compiler failing. These  were not fatal, but the
script would end up running on the CPUs rather than the GPU.

 W Adreno-RS: <rsdVendorScriptInitQCOM:795>: ERROR: rsdCompileBitcode returned -30

Change-Id: I8c58900c3268b6998d961cd6c9e0c09b0d9971f6